+++ src/test/java/racingcar/domain/RacingResultTest.java
@@ -7,6 +7,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.service.dto.CarDto;
+
 public class RacingResultTest {
 
 	@Test
